pyproject.toml
[project]
name = "openbrain"
version = "0.1.0"
description = "Open Brain: AI-native second brain MCP server"
requires-python = ">=3.12"

dependencies = [
    "fastmcp>=2.3.0",
    "azure-cosmos>=4.9.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.10.0",
    "python-dotenv>=1.0.1",
    "openai>=1.60.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/openbrain"]
